Set up all of the methods to restful
Important: I added another interface: game.get, which is called when the game ends.

This is seperate from the other "game" stuff because of the way the frontend is coded.

The frontend needs to decide when "game ends" happens because it needs to stop the timer and change the UI. This can't be done when the page is posted because of the setup :c.

We can look at changing this in the future, but will need to communicate with frontend people.

I also reformatted a LOT of the network emssage class, sends back JSON objects instead now, which are parsed in the netwroker.